CLI: Download Detector Configuration as File (#229)
* Refactor cat command

Go provides an abstraction to support writing on file
and stdout as same interface, refactor cat to accept
func as input. Later, download will pass writeInfile as parameter

* CLI: Refactor command to display error

Printing error on console is added everywhere.
Create a function to check error and print.

* Add Download command

Download is similar to cat, excepts it download the configuration and
save it as file with detector name as file name on current folder

* CLI: Add interactive to prompt for overwriting files

Similar to cp from linux, add interactive to allow user to decide
whether to overwrite file or not

* CLI: Allow user to specify output folder path

Added flag to allow user to specify target location where
detectors will be downloaded.

* CLI: Make download prompt as default

Download detector will always prompt if there is file
exists in output directory.